+++ cobaya/component.py
@@ -6,6 +6,7 @@ from cobaya.log import HasLogger, LoggedError
 from cobaya.input import HasDefaults
 from cobaya.typing import InfoDict, InfoDictIn, empty_dict
 from cobaya.tools import resolve_packages_path
+from cobaya.conventions import packages_path_input
 
 
 class Timer:
@@ -77,13 +78,14 @@ class CobayaComponent(HasLogger, HasDefaults):
         for k, value in info.items():
             try:
                 # MARKED FOR DEPRECATION IN v3.0
+                # NB: cannot ever raise an error, since users may use "path_install" for
+                #     their own purposes. When considered *fully* deprecated, simply
+                #     remove this whole block.
                 if k == "path_install":
                     self.log.warning(
                         "*DEPRECATION*: `path_install` will be deprecated "
                         "in the next version. Please use `packages_path` instead.")
-                    # BEHAVIOUR TO BE REPLACED BY ERROR:
-                    # set BOTH old and new names, just in case old one is used internally
-                    setattr(self, "packages_path", value)
+                    setattr(self, packages_path_input, value)
                 # END OF DEPRECATION BLOCK
                 setattr(self, k, value)
             except AttributeError: