+++ cobaya/likelihoods/base_classes/bao.py
@@ -2,7 +2,8 @@ r"""
 .. module:: bao
 
 :Synopsis: BAO, f_sigma8 and other measurements at single redshifts, with correlations
-:Author: Antony Lewis (adapted to Cobaya by Jesus Torrado, with little modification)
+:Author: Antony Lewis, Pablo Lemos (adapted to Cobaya by Jesus Torrado, with little
+        modification)
 
 This code provides a template for BAO, :math:`f\sigma_8`, :math:`H`
 and other redshift dependent functions.
@@ -14,6 +15,10 @@ The datasets implemented at this moment are:
 - ``bao.sdss_dr12_consensus_bao``
 - ``bao.sdss_dr12_consensus_full_shape``
 - ``bao.sdss_dr12_consensus_final`` (combined data of the previous two)
+- ``bao.sdss_dr16_baoplus_lrg`` (combining data from BOSS DR12 and eBOSS DR16)
+- ``bao.sdss_dr16_baoplus_lyauto``
+- ``bao.sdss_dr16_baoplus_lyxqso``
+- ``bao.sdss_dr16_baoplus_qso``
 
 .. |br| raw:: html
 
@@ -40,6 +45,13 @@ The datasets implemented at this moment are:
    `The clustering of galaxies in the completed SDSS-III Baryon Oscillation Spectroscopic
    Survey: cosmological analysis of the DR12 galaxy sample`
    `(arXiv:1607.03155) <https://arxiv.org/abs/1607.03155>`_
+   |br| |br|
+   If you use any of the likelihoods ``sdss_dr16_*``, please cite:
+   |br|
+   S. Alam et al,
+   `The Completed SDSS-IV extended Baryon Oscillation Spectroscopic Survey: Cosmological
+   Implications from two Decades of Spectroscopic Surveys at the Apache Point observatory`
+   `(arXiv:2007.08991) <https://arxiv.org/abs/2007.08991>`_
 
 
 Usage
@@ -114,21 +126,18 @@ After this, mention the path to this likelihood when you include it in an input 
    likelihood:
      bao.sdss_dr12_consensus_[bao|full_shape|final|...]:
        path: /path/to/likelihoods/bao_data
-
-
 """
-
 # Global
 import os
 import numpy as np
-from scipy.interpolate import UnivariateSpline
+from scipy.interpolate import UnivariateSpline, RectBivariateSpline, \
+    RegularGridInterpolator
 import pandas as pd
 from typing import Optional, Sequence
 
 # Local
 from cobaya.log import LoggedError
-from cobaya.conventions import Const
-from cobaya.typing import InfoDict
+from cobaya.conventions import Const, packages_path_input
 from cobaya.likelihoods.base_classes import InstallableLikelihood
 
 
@@ -137,7 +146,7 @@ class BAO(InstallableLikelihood):
     type = "BAO"
 
     install_options = {"github_repository": "CobayaSampler/bao_data",
-                       "github_release": "v1.1"}
+                       "github_release": "v2.0"}
 
     prob_dist_bounds: Optional[Sequence[float]]
     measurements_file: Optional[str] = None
@@ -146,14 +155,20 @@ class BAO(InstallableLikelihood):
     prob_dist: Optional[str] = None
     cov_file: Optional[str] = None
     invcov_file: Optional[str] = None
+    redshift: Optional[float] = None
+    observable_1: Optional[str] = None
+    observable_2: Optional[str] = None
+    observable_3: Optional[str] = None
+    grid_file: Optional[str] = None
     path: Optional[str]
 
     def initialize(self):
         self.log.info("Initialising.")
-        if not getattr(self, "path", None) and not getattr(self, "packages_path", None):
+        if not getattr(self, "path", None) and \
+                not getattr(self, packages_path_input, None):
             raise LoggedError(
                 self.log, "No path given to BAO data. Set the likelihood property "
-                          "'path' or the common property '%s'.", "packages_path")
+                          "'path' or the common property '%s'.", packages_path_input)
         # If no path specified, use the external packages path
         data_file_path = os.path.normpath(getattr(self, "path", None) or
                                           os.path.join(self.packages_path, "data"))
@@ -173,20 +188,27 @@ class BAO(InstallableLikelihood):
                 raise LoggedError(
                     self.log, "Couldn't find measurements file '%s' in folder '%s'. " % (
                         self.measurements_file, data_file_path) + "Check your paths.")
+        elif self.grid_file:
+            pass
         else:
             self.data = pd.DataFrame([self.data] if not hasattr(self.data[0], "__len__")
                                      else self.data)
-        # Columns: z value [err] [type]
-        self.has_type = self.data.iloc[:, -1].dtype == np.dtype("O")
-        assert self.has_type  # mandatory for now!
-        self.has_err = len(self.data.columns) > 2 and self.data[2].dtype == float
-        if self.has_err:
-            self.data.columns = ["z", "value", "error", "observable"]
-        else:
-            self.data.columns = ["z", "value", "observable"]
-        prefix = "bao_"
-        self.data["observable"] = [(c[len(prefix):] if c.startswith(prefix) else c)
-                                   for c in self.data["observable"]]
+
+        if not self.grid_file:
+            self.use_grid_2d = False
+            self.use_grid_3d = False
+            # Columns: z value [err] [type]
+            self.has_type = self.data.iloc[:, -1].dtype == np.dtype("O")
+            assert self.has_type  # mandatory for now!
+            self.has_err = len(self.data.columns) > 2 and self.data[2].dtype == float
+            if self.has_err:
+                self.data.columns = ["z", "value", "error", "observable"]
+            else:
+                self.data.columns = ["z", "value", "observable"]
+            prefix = "bao_"
+            self.data["observable"] = [(c[len(prefix):] if c.startswith(prefix) else c)
+                                       for c in self.data["observable"]]
+
         # Probability distribution
         if self.prob_dist:
             try:
@@ -205,11 +227,75 @@ class BAO(InstallableLikelihood):
                     self.log, "If 'prob_dist' given, 'prob_dist_bounds' needs to be "
                               "specified as [min, max].")
             spline = UnivariateSpline(alpha, -chi2 / 2, s=0)
-            self.logpdf = lambda x: (
-                spline(x)[0] if self.prob_dist_bounds[0] <= x <= self.prob_dist_bounds[1]
-                else -np.inf)
-        # Covariance --> read and re-sort as self.data
+            self.logpdf = lambda _x: (spline(_x)[0] if self.prob_dist_bounds[0]
+                                                       <= _x <= self.prob_dist_bounds[1]
+                                      else -np.inf)
+        elif self.grid_file:
+            try:
+                self.grid_data = np.loadtxt(
+                    os.path.join(data_file_path, self.grid_file))
+            except IOError:
+                raise LoggedError(
+                    self.log, "Couldn't find grid file '%s' in folder '%s'. " % (
+                        self.grid_file, data_file_path) + "Check your paths.")
+            if not self.observable_1 or not self.observable_2:
+                raise LoggedError(
+                    self.log, "If using grid data, 'observable_1' and 'observable_2'"
+                              "need to be specified.")
+            if self.redshift is None:
+                raise LoggedError(
+                    self.log, "If using grid data, 'redshift'"
+                              "needs to be specified.")
+
+            self.has_type = True  # Not sure what this is
+            self.data = pd.DataFrame()
+
+            if self.grid_data.shape[1] == 3:
+                self.use_grid_2d = True
+                self.use_grid_3d = False
+                self.data["observable"] = [self.observable_1, self.observable_2]
+
+                x = np.unique(self.grid_data[:, 0])
+                y = np.unique(self.grid_data[:, 1])
+
+                Nx = x.shape[0]
+                Ny = y.shape[0]
+
+                chi2 = np.reshape(np.log(self.grid_data[:, 2]), [Nx, Ny])
+
+                # Make the interpolator (x refers to at, y refers to ap).
+                self.interpolator = RectBivariateSpline(x, y, chi2, kx=3, ky=3)
+            elif self.grid_data.shape[1] == 4:
+                self.use_grid_2d = False
+                self.use_grid_3d = True
+                if not self.observable_3:
+                    raise LoggedError(
+                        self.log, "If using 3D grid data, 'observable_3'"
+                                  "needs to be specified.")
+                self.data["observable"] = [self.observable_1, self.observable_2,
+                                           self.observable_3]
+
+                x = np.unique(self.grid_data[:, 0])
+                y = np.unique(self.grid_data[:, 1])
+                z = np.unique(self.grid_data[:, 2])
+
+                Nx = x.shape[0]
+                Ny = y.shape[0]
+                Nz = z.shape[0]
+
+                chi2 = np.reshape(np.log(self.grid_data[:, 3] + 1e-300), [Nx, Ny, Nz])
+
+                self.interpolator3D = RegularGridInterpolator((x, y, z), chi2,
+                                                              bounds_error=False,
+                                                              fill_value=np.log(1e-300))
+
+            else:
+                raise LoggedError(
+                    self.log, "Grid data has the wrong dimensions")
+                # Covariance --> read and re-sort as self.data
         else:
+            self.use_grid_2d = False
+            self.use_grid_3d = False
             try:
                 if self.cov_file:
                     self.cov = np.loadtxt(os.path.join(data_file_path, self.cov_file))
@@ -228,14 +314,23 @@ class BAO(InstallableLikelihood):
                     self.log, "Couldn't find (inv)cov file '%s' in folder '%s'. " % (
                         self.cov_file or self.invcov_file,
                         data_file_path) + "Check your paths.")
-            self.logpdf = lambda x: (lambda x_: -0.5 * x_.dot(self.invcov).dot(x_))(
-                x - self.data["value"].values)
+            self.logpdf = lambda _x: (lambda x_: -0.5 * x_.dot(self.invcov).dot(x_))(
+                _x - self.data["value"].values)
 
     def get_requirements(self):
         # Requisites
-        zs = {obs: self.data.loc[self.data["observable"] == obs, "z"].values
-              for obs in self.data["observable"].unique()}
-        theory_reqs: InfoDict = {
+        if self.use_grid_2d:
+            zs = {self.observable_1: np.array([self.redshift]),
+                  self.observable_2: np.array([self.redshift])}
+        elif self.use_grid_3d:
+            zs = {self.observable_1: np.array([self.redshift]),
+                  self.observable_2: np.array([self.redshift]),
+                  self.observable_3: np.array([self.redshift])
+                  }
+        else:
+            zs = {obs: self.data.loc[self.data["observable"] == obs, "z"].values
+                  for obs in self.data["observable"].unique()}
+        theory_reqs = {
             "DV_over_rs": {
                 "angular_diameter_distance": {"z": zs.get("DV_over_rs", None)},
                 "Hubble": {"z": zs.get("DV_over_rs", None)},
@@ -258,7 +353,7 @@ class BAO(InstallableLikelihood):
                 "rdrag": None},
             "f_sigma8": {
                 "fsigma8": {"z": zs.get("f_sigma8", None)},
-                },
+            },
             "F_AP": {
                 "angular_diameter_distance": {"z": zs.get("F_AP", None)},
                 "Hubble": {"z": zs.get("F_AP", None)}}}
@@ -314,11 +409,23 @@ class BAO(InstallableLikelihood):
         return self.provider.get_param("rdrag") * self.rs_rescale
 
     def logp(self, **params_values):
-        theory = np.array([self.theory_fun(z, obs) for z, obs
-                           in zip(self.data["z"], self.data["observable"])]).T[0]
-        if self.is_debug():
-            for i, (z, obs, theo) in enumerate(
-                    zip(self.data["z"], self.data["observable"], theory)):
-                self.log.debug("%s at z=%g : %g (theo) ; %g (data)",
-                               obs, z, theo, self.data.iloc[i, 1])
-        return self.logpdf(theory)
+        if self.use_grid_2d:
+            x = self.theory_fun(self.redshift, self.observable_1)
+            y = self.theory_fun(self.redshift, self.observable_2)
+            chi2 = float(self.interpolator(x, y)[0])
+            return chi2 / 2
+        elif self.use_grid_3d:
+            x = self.theory_fun(self.redshift, self.observable_1)
+            y = self.theory_fun(self.redshift, self.observable_2)
+            z = self.theory_fun(self.redshift, self.observable_3)
+            chi2 = self.interpolator3D(np.array([x, y, z])[:, 0])
+            return chi2 / 2
+        else:
+            theory = np.array([self.theory_fun(z, obs) for z, obs
+                               in zip(self.data["z"], self.data["observable"])]).T[0]
+            if self.is_debug():
+                for i, (z, obs, theo) in enumerate(
+                        zip(self.data["z"], self.data["observable"], theory)):
+                    self.log.debug("%s at z=%g : %g (theo) ; %g (data)",
+                                   obs, z, theo, self.data.iloc[i, 1])
+            return self.logpdf(theory)