+++ cobaya/post.py
@@ -16,7 +16,8 @@ import numpy as np
 from cobaya import mpi
 from cobaya.collection import SampleCollection
 from cobaya.conventions import prior_1d_name, OutPar, get_chi2_name, \
-    undo_chi2_name, get_minuslogpior_name, separator_files, minuslogprior_names
+    undo_chi2_name, get_minuslogpior_name, separator_files, minuslogprior_names, \
+    packages_path_input
 from cobaya.input import update_info, add_aggregated_chi2_params, load_input_dict
 from cobaya.log import logger_setup, get_logger, is_debug, LoggedError
 from cobaya.model import Model
@@ -26,7 +27,7 @@ from cobaya.parameterization import is_fixed_or_function_param, is_sampled_param
     is_derived_param
 from cobaya.prior import Prior
 from cobaya.tools import progress_bar, recursive_update, deepcopy_where_possible, \
-    check_deprecated_modules_path, str_to_list
+    str_to_list
 from cobaya.typing import ExpandedParamsDict, ModelBlock, ParamValuesDict, InputDict, \
     InfoDict, PostDict
 
@@ -83,8 +84,9 @@ def post(info_or_yaml_or_file: Union[InputDict, str, os.PathLike],
     logger_setup(info.get("debug"), info.get("debug_file"))
     log = get_logger(__name__)
     # MARKED FOR DEPRECATION IN v3.0
-    # BEHAVIOUR TO BE REPLACED BY ERROR:
-    check_deprecated_modules_path(info)
+    if info.get("modules"):
+        raise LoggedError(log, "The input field 'modules' has been deprecated."
+                               "Please use instead %r", packages_path_input)
     # END OF DEPRECATION BLOCK
     info_post: PostDict = info.get("post") or {}
     if not info_post:
@@ -342,8 +344,8 @@ def post(info_or_yaml_or_file: Union[InputDict, str, os.PathLike],
     # TODO: check allow_renames=False?
     model_add = Model(out_params_with_computed, add["likelihood"],
                       info_prior=add.get("prior"), info_theory=out_combined["theory"],
-                      packages_path=(info_post.get("packages_path") or
-                                     info.get("packages_path")),
+                      packages_path=(info_post.get(packages_path_input) or
+                                     info.get(packages_path_input)),
                       allow_renames=False, post=True,
                       stop_at_error=info.get('stop_at_error', False),
                       skip_unused_theories=True, dropped_theory_params=dropped_theory)
@@ -407,7 +409,8 @@ def post(info_or_yaml_or_file: Union[InputDict, str, os.PathLike],
 
         def set_difflogmax():
             nonlocal difflogmax
-            difflog = (collection_in[OutPar.minuslogpost].to_numpy(dtype=np.float64)[:len(collection_out)]
+            difflog = (collection_in[OutPar.minuslogpost].to_numpy(
+                dtype=np.float64)[:len(collection_out)]
                        - collection_out[OutPar.minuslogpost].to_numpy(dtype=np.float64))
             difflogmax = np.max(difflog)
             if abs(difflogmax) < 1:
@@ -456,7 +459,8 @@ def post(info_or_yaml_or_file: Union[InputDict, str, os.PathLike],
             if -np.inf in logpriors_new:
                 continue
             # Add/remove likelihoods and/or (re-)calculate derived parameters
-            loglikes_add, output_derived = model_add.logps(all_params)
+            loglikes_add, output_derived = model_add._loglikes_input_params(
+                all_params, return_output_params=True)
             loglikes_add = dict(zip(chi2_names_add, loglikes_add))
             output_derived = dict(zip(model_add.output_params, output_derived))
             loglikes_new = [loglikes_add.get(name, -0.5 * point.get(name, 0))
@@ -471,8 +475,8 @@ def post(info_or_yaml_or_file: Union[InputDict, str, os.PathLike],
             all_params.update(output_derived)
 
             all_params.update(out_func_parameterization.to_derived(all_params))
-            derived = {param: all_params.get(param)
-                       for param in dummy_model_out.parameterization.derived_params()}
+            derived = {param: all_params.get(param) for param in
+                       dummy_model_out.parameterization.derived_params()}
             # We need to recompute the aggregated chi2 by hand
             for type_, likes in inv_types.items():
                 derived[get_chi2_name(type_)] = sum(