+++ cobaya/theories/camb/camb.py
@@ -283,6 +283,7 @@ class CAMB(BoltzmannBase):
         if not self.external_primordial_pk \
                 and set(self.input_params).intersection({'r', 'At'}):
             self.extra_attrs["WantTensors"] = True
+            self.extra_attrs["Accuracy.AccurateBB"] = True
 
     def get_can_support_params(self):
         return self.power_params + self.nonlin_params
@@ -292,7 +293,7 @@ class CAMB(BoltzmannBase):
 
     def set_cl_reqs(self, reqs):
         """
-        Sets some common settings for both lensend and unlensed Cl's.
+        Sets some common settings for both lensed and unlensed Cl's.
         """
         self.extra_args["lmax"] = max(
             max(reqs.values()), self.extra_args.get("lmax", 0))
@@ -517,8 +518,6 @@ class CAMB(BoltzmannBase):
                     args.update(self.nonlin_args)
                     results.Params.NonLinearModel.set_params(**args)
                 results.power_spectra_from_transfer()
-            else:
-                results = None
             for product, collector in self.collectors.items():
                 if collector:
                     state[product] = \
@@ -726,8 +725,16 @@ class CAMB(BoltzmannBase):
                     self.log.debug("Setting attributes of CAMBparams: %r",
                                    self.extra_attrs)
                 for attr, value in self.extra_attrs.items():
-                    if hasattr(params, attr):
-                        setattr(params, attr, value)
+                    obj = params
+                    if '.' in attr:
+                        parts = attr.split('.')
+                        for p in parts[:-1]:
+                            obj = getattr(obj, p)
+                        par = parts[-1]
+                    else:
+                        par = attr
+                    if hasattr(obj, par):
+                        setattr(obj, par, value)
                     else:
                         raise LoggedError(
                             self.log,