+++ cobaya/tools.py
@@ -30,7 +30,8 @@ import traceback
 
 # Local
 from cobaya.conventions import cobaya_package, subfolders, kinds, \
-    packages_path_config_file, packages_path_env, packages_path_arg, dump_sort_cosmetic
+    packages_path_config_file, packages_path_env, packages_path_arg, dump_sort_cosmetic, \
+    packages_path_input
 from cobaya.log import LoggedError, HasLogger, get_logger
 from cobaya.typing import Kind
 
@@ -998,7 +999,7 @@ def load_packages_path_from_config_file():
     Returns the external packages path stored in the config file,
     or `None` if it can't be found.
     """
-    return load_config_file().get("packages_path")
+    return load_config_file().get(packages_path_input)
 
 
 def write_packages_path_in_config_file(packages_path):
@@ -1007,7 +1008,7 @@ def write_packages_path_in_config_file(packages_path):
 
     Relative paths are converted into absolute ones.
     """
-    write_config_file({"packages_path": os.path.abspath(packages_path)})
+    write_config_file({packages_path_input: os.path.abspath(packages_path)})
 
 
 def resolve_packages_path(infos=None):
@@ -1028,10 +1029,13 @@ def resolve_packages_path(infos=None):
     elif isinstance(infos, Mapping):
         infos = [infos]
     # MARKED FOR DEPRECATION IN v3.0
-    # BEHAVIOUR TO BE REPLACED BY ERROR:
-    [check_deprecated_modules_path(info) for info in infos]
+    for info in infos:
+        if info.get("modules"):
+            raise LoggedError(log, "The input field 'modules' has been deprecated."
+                                   "Please use instead %r", packages_path_input)
     # END OF DEPRECATION BLOCK
-    paths = set(p for p in [info.get("packages_path") for info in infos] if p)
+    paths = set(os.path.realpath(p) for p in
+                [info.get(packages_path_input) for info in infos] if p)
     if len(paths) == 1:
         return list(paths)[0]
     elif len(paths) > 1:
@@ -1045,12 +1049,9 @@ def resolve_packages_path(infos=None):
     old_env = "COBAYA_MODULES"
     path_old_env = os.environ.get(old_env)
     if path_old_env and not path_env:
-        log.warning("*DEPRECATION*: The env var %r will be deprecated in favor of %r in "
-                    "the next version. Please, use that one instead.",
-                    old_env, packages_path_env)
-        # BEHAVIOUR TO BE REPLACED BY ERROR:
-        path_env = path_old_env
-    # END OF DEPRECATION BLOCK -- CONTINUES BELOW!
+        raise LoggedError(log, "The env var %r has been deprecated in favor of %r",
+                          old_env, packages_path_env)
+    # END OF DEPRECATION BLOCK
     if path_env:
         return path_env
     return load_packages_path_from_config_file()
@@ -1059,24 +1060,12 @@ def resolve_packages_path(infos=None):
 def sort_cosmetic(info):
     # noinspection PyStatementEffect
     """
-        Returns a sorted version of the given info dict, re-ordered as %r, and finally the
-        rest of the blocks/options.
-        """ % dump_sort_cosmetic
+    Returns a sorted version of the given info dict, re-ordered as %r, and finally the
+    rest of the blocks/options.
+    """ % dump_sort_cosmetic
     sorted_info = dict()
     for k in dump_sort_cosmetic:
         if k in info:
             sorted_info[k] = info[k]
     sorted_info.update({k: v for k, v in info.items() if k not in sorted_info})
     return sorted_info
-
-
-# MARKED FOR DEPRECATION IN v3.0
-def check_deprecated_modules_path(info):
-    if info.get("modules"):
-        log.warning("*DEPRECATION*: The input field 'modules' will be deprecated in "
-                    "favor of %r in the next version. Please, use that one instead.",
-                    "packages_path")
-        # BEHAVIOUR TO BE REPLACED BY ERROR:
-        if not info.get("packages_path"):
-            info["packages_path"] = info["modules"]
-# END OF DEPRECATION BLOCK